✨ (server): add custom resolver for Mission.missionPatch
The resolver for Mission.missionPatch should return a different value depending on whether a query specifies LARGE or SMALL for the size argument.
This resolver obtains a large or small patch from mission, which is the object returned by the default resolver for the parent field in our schema, Launch.mission.