✨ (web): add pagination support
Apollo Client provides a fetchMore helper function to assist with paginated queries.
It enables to execute the same query with different values for variables (such as the current cursor).
Apollo Client stores query results in its in-memory cache.
The cache handles most operations intelligently and efficiently, but it doesn't automatically know that we want to merge our two distinct lists of launches.
To fix this, we define a merge function for the paginated field in our schema.